+++ R/TBox-package.R
@@ -7,7 +7,12 @@
 #' @importFrom clipr read_clip
 #' @importFrom clipr write_clip
 #' @importFrom knitr opts_chunk
+#' @importFrom rmarkdown html_document
+#' @importFrom rmarkdown pdf_document
 #' @importFrom rmarkdown render
+#' @importFrom rmarkdown word_document
 #' @importFrom utils browseURL
+#' @importFrom utils read.csv
+#' @importFrom utils write.table
 ## usethis namespace: end
 NULL