+++ tests/testthat.R
@@ -6,7 +6,7 @@
 # * https://r-pkgs.org/testing-design.html#sec-tests-files-overview
 # * https://testthat.r-lib.org/articles/special-files.html
 
-library(testthat)
-library(TBox)
+library("testthat") # nolint unused_import_linter
+library("TBox") # nolint unused_import_linter
 
 test_check("TBox")