+++ tests/testthat/helper.R
@@ -1,0 +1,2 @@
+
+latex_engine <- get_latex_engine()