+++ tests/testthat/test-generate_chunk_header.R
@@ -22,8 +22,8 @@ testthat::test_that("The header is well generated", {
                           "comment = \"##\", background = \"#E7E7E7\"}")
     )
     testthat::expect_identical(
-        object = generate_chunk_header(fig.width = 8.2, fig.height = 10,
-                                       dpi = 124, warning = FALSE),
+        object = generate_chunk_header(fig.width = 8.2, fig.height = 10.,
+                                       dpi = 124., warning = FALSE),
         expected = paste0("```{r, fig.width = 8.2, fig.height = 10, ",
                           "dpi = 124, warning = FALSE}")
     )
@@ -32,8 +32,10 @@ testthat::test_that("The header is well generated", {
 ## Tests qui échouent
 
 testthat::test_that("The header is well generated", {
-    testthat::expect_error(object = generate_chunk_header(eval = TRUE,
-                                                          eval = FALSE))
+    testthat::expect_error(object = generate_chunk_header(
+        eval = TRUE,
+                                                          eval = FALSE # nolint undesirable_function_linter
+    ))
     testthat::expect_error(object = generate_chunk_header(Eval = FALSE))
     testthat::expect_error(object = generate_chunk_header(result = "markup"))
 })