+++ tests/testthat/test-generate_rmd_file.R
@@ -1,113 +1,148 @@
 
 testthat::test_that("Everything works with default parameters", {
-    testthat::expect_identical(
-        object = generate_rmd_file(output = "word",
-                                   content = "plot(AirPassengers)"),
-        expected = paste0("---\ntitle: \"Format code\"\noutput:\n  ",
-                          "word_document:\n    highlight: arrow\n",
-                          "monofont: \"Fira Code\"\ncode-block-bg: ",
-                          "true\ncode-block-border-left: \"#31BAE9\"\n",
-                          "---\n\n## Running Code\n\n```{r}\n",
-                          "plot(AirPassengers)\n```\n")
+
+    exp1 <- readLines(testthat::test_path("template_rmd", "default",
+                                          "template_word.Rmd"))
+    obj1_a <- generate_rmd_file(
+        output_format = "word",
+        content = "plot(AirPassengers)"
     )
-    testthat::expect_identical(
-        object = generate_rmd_file(output = "html",
-                                   content = "plot(AirPassengers)"),
-        expected = paste0("---\ntitle: \"Format code\"\noutput:\n  ",
-                          "html_document:\n    highlight: arrow\nmonofont",
-                          ": \"Fira Code\"\ncode-block-bg: true\n",
-                          "code-block-border-left: \"#31BAE9\"\n---\n\n",
-                          "## Running Code\n\n```{r}\nplot(AirPassengers)",
-                          "\n```\n")
+    obj1_b <- generate_rmd_file(
+        output_format = "word_document",
+        content = "plot(AirPassengers)"
     )
-    testthat::expect_identical(
-        object = generate_rmd_file(output = "pdf",
-                                   content = "plot(AirPassengers)",
-                                   fira_path = "User"),
-        expected = paste0("---\ntitle: \"Format code\"\noutput:\n  ",
-                          "pdf_document:\n    highlight: arrow\n    ",
-                          "latex_engine: xelatex\ncode-block-bg: true\n",
-                          "code-block-border-left: \"#31BAE9\"\n---\n\n",
-                          "\\fontsize{12}{12}\n\\setmonofont[",
-                          "ExternalLocation=User/]{FiraCode-Regular.ttf}",
-                          "\n\n## Running Code\n\n```{r}\n",
-                          "plot(AirPassengers)\n```\n")
+
+    testthat::expect_identical(object = obj1_a, expected = exp1)
+    testthat::expect_identical(object = obj1_b, expected = exp1)
+
+
+    exp2 <- readLines(testthat::test_path("template_rmd", "default",
+                                          "template_html.Rmd"))
+    obj2_a <- generate_rmd_file(output_format = "html",
+                                content = "plot(AirPassengers)")
+    obj2_b <- generate_rmd_file(output_format = "html_document",
+                                content = "plot(AirPassengers)")
+
+    testthat::expect_identical(object = obj2_a, expected = exp2)
+    testthat::expect_identical(object = obj2_b, expected = exp2)
+
+
+    exp3 <- readLines(testthat::test_path(
+        "template_rmd",
+        "default",
+        paste0("template_pdf.Rmd")
+    ))
+    obj3_a <- generate_rmd_file(
+        output_format = "pdf",
+        content = "plot(AirPassengers)"
     )
+    obj3_b <- generate_rmd_file(
+        output_format = "pdf_document",
+        content = "plot(AirPassengers)"
+    )
+
+    testthat::expect_identical(object = obj3_a, expected = exp3)
+    testthat::expect_identical(object = obj3_b, expected = exp3)
 })
 
 testthat::test_that("Everything works with custom parameters", {
-    testthat::expect_identical(
-        object = generate_rmd_file(output = "word",
-                                   content = "Bonjour tout le monde",
-                                   code = FALSE),
-        expected = paste0("---\n",
-                          "title: \"Format code\"\n",
-                          "output:\n",
-                          "  word_document:\n",
-                          "    highlight: arrow\n",
-                          "monofont: \"Fira Code\"\n",
-                          "code-block-bg: true\n",
-                          "code-block-border-left: \"#31BAE9\"\n",
-                          "---\n\n",
-                          "## Running Code\n\n\n",
-                          "Bonjour tout le monde\n\n")
+
+    exp1 <- readLines(testthat::test_path("template_rmd", "custom",
+                                          "template_word_1.Rmd"))
+    obj1_a <- generate_rmd_file(
+        output_format = "word",
+        content = "Bonjour tout le monde",
+        code = FALSE
     )
-    testthat::expect_identical(
-        object = generate_rmd_file(output = "html",
-                                   content = "Bonjour tout le monde",
-                                   code = FALSE),
-        expected = paste0("---\ntitle: \"Format code\"\noutput:",
-                          "\n  html_document:\n    highlight: arrow\n",
-                          "monofont: \"Fira Code\"\ncode-block-bg: true\n",
-                          "code-block-border-left: \"#31BAE9\"\n---\n\n",
-                          "## Running Code\n\n\nBonjour tout le monde\n\n")
+    obj1_b <- generate_rmd_file(
+        output_format = "word_document",
+        content = "Bonjour tout le monde",
+        code = FALSE
     )
-    testthat::expect_identical(
-        object = generate_rmd_file(output = "pdf",
-                                   content = "Bonjour tout le monde",
-                                   code = FALSE, fira_path = "User"),
-        expected = paste0("---\ntitle: \"Format code\"\noutput:\n",
-                          "  pdf_document:\n    highlight: arrow\n",
-                          "    latex_engine: xelatex\n",
-                          "code-block-bg: true\n",
-                          "code-block-border-left: \"#31BAE9\"\n---\n\n",
-                          "\\fontsize{12}{12}\n\\setmonofont[",
-                          "ExternalLocation=User/]{FiraCode-Regular.ttf}\n\n",
-                          "## Running Code\n\n\nBonjour tout le monde\n\n")
+
+    testthat::expect_identical(object = obj1_a, expected = exp1)
+    testthat::expect_identical(object = obj1_b, expected = exp1)
+
+
+    exp2 <- readLines(testthat::test_path("template_rmd", "custom",
+                                          "template_html_1.Rmd"))
+    obj2_a <- generate_rmd_file(output_format = "html",
+                                content = "Bonjour tout le monde",
+                                code = FALSE)
+    obj2_b <- generate_rmd_file(output_format = "html_document",
+                                content = "Bonjour tout le monde",
+                                code = FALSE)
+
+    testthat::expect_identical(object = obj2_a, expected = exp2)
+    testthat::expect_identical(object = obj2_b, expected = exp2)
+
+
+    exp3 <- readLines(testthat::test_path(
+        "template_rmd",
+        "custom",
+        paste0("template_pdf_1.Rmd")
+    ))
+    obj3_a <- generate_rmd_file(
+        output_format = "pdf",
+        content = "Bonjour tout le monde",
+        code = FALSE
     )
-    testthat::expect_identical(
-        object = generate_rmd_file(output = "word",
-                                   content = "plot(AirPassengers)",
-                                   eval = TRUE),
-        expected = paste0("---\ntitle: \"Format code\"\noutput:\n",
-                          "  word_document:\n    highlight: arrow\n",
-                          "monofont: \"Fira Code\"\ncode-block-bg: true\n",
-                          "code-block-border-left: \"#31BAE9\"\n---\n\n",
-                          "## Running Code\n\n```{r, eval = TRUE}\n",
-                          "plot(AirPassengers)\n```\n")
+    obj3_b <- generate_rmd_file(
+        output_format = "pdf_document",
+        content = "Bonjour tout le monde",
+        code = FALSE
     )
-    testthat::expect_identical(
-        object = generate_rmd_file(output = "html",
-                                   content = "plot(AirPassengers)",
-                                   eval = TRUE),
-        expected = paste0("---\ntitle: \"Format code\"\noutput:\n",
-                          "  html_document:\n    highlight: arrow\n",
-                          "monofont: \"Fira Code\"\ncode-block-bg: true\n",
-                          "code-block-border-left: \"#31BAE9\"\n---\n\n",
-                          "## Running Code\n\n```{r, eval = TRUE}\n",
-                          "plot(AirPassengers)\n```\n")
+
+    testthat::expect_identical(object = obj3_a, expected = exp3)
+    testthat::expect_identical(object = obj3_b, expected = exp3)
+
+
+    exp4 <- readLines(testthat::test_path("template_rmd", "custom",
+                                          "template_word_2.Rmd"))
+    obj4_a <- generate_rmd_file(
+        output_format = "word",
+        content = "plot(AirPassengers)",
+        eval = TRUE
     )
-    testthat::expect_identical(
-        object = generate_rmd_file(output = "pdf",
-                                   content = "plot(AirPassengers)",
-                                   eval = TRUE, fira_path = "User"),
-        expected = paste0("---\ntitle: \"Format code\"\noutput:\n",
-                          "  pdf_document:\n    highlight: arrow\n",
-                          "    latex_engine: xelatex\ncode-block-bg: true\n",
-                          "code-block-border-left: \"#31BAE9\"\n---\n\n",
-                          "\\fontsize{12}{12}\n\\setmonofont[",
-                          "ExternalLocation=User/]{FiraCode-Regular.ttf}\n\n",
-                          "## Running Code\n\n```{r, eval = TRUE}\n",
-                          "plot(AirPassengers)\n```\n")
+    obj4_b <- generate_rmd_file(
+        output_format = "word_document",
+        content = "plot(AirPassengers)",
+        eval = TRUE
     )
+
+    testthat::expect_identical(object = obj4_a, expected = exp4)
+    testthat::expect_identical(object = obj4_b, expected = exp4)
+
+
+    exp5 <- readLines(testthat::test_path("template_rmd", "custom",
+                                          "template_html_2.Rmd"))
+    obj5_a <- generate_rmd_file(output_format = "html",
+                                content = "plot(AirPassengers)",
+                                eval = TRUE)
+    obj5_b <- generate_rmd_file(output_format = "html_document",
+                                content = "plot(AirPassengers)",
+                                eval = TRUE)
+
+    testthat::expect_identical(object = obj5_a, expected = exp5)
+    testthat::expect_identical(object = obj5_b, expected = exp5)
+
+
+    exp6 <- readLines(testthat::test_path(
+        "template_rmd",
+        "custom",
+        paste0("template_pdf_2.Rmd")
+    ))
+    obj6_a <- generate_rmd_file(
+        output_format = "pdf",
+        content = "plot(AirPassengers)",
+        eval = TRUE
+    )
+    obj6_b <- generate_rmd_file(
+        output_format = "pdf_document",
+        content = "plot(AirPassengers)",
+        eval = TRUE
+    )
+
+    testthat::expect_identical(object = obj6_a, expected = exp6)
+    testthat::expect_identical(object = obj6_b, expected = exp6)
 })